Restrict scanned chunks for certain limit queries
When limit queries use 'order by' and the first attribute of the order
by is equivalent to the first attribute of the hypertable it can be
beneficial to sort chunks individually. In that case ChunkAppend can be
used and only the needed chunks are accessed (e.g., if the query has a
small limit value, only the first chunk needs to be accessed in most
cases).

If at least one chunk of a hypertable is uncompressed, PostgreSQL will
generate a MergeAppend path in generate_orderedappend_paths() /
create_merge_append_path() due to the existing pathkeys of the index on
the uncompressed chunk. If no path keys are present no MergeAppend path
is generated by PostgreSQL.  As a result, an expensive (parallel) query
plan with the sort node on top of the DecompressChunkNodes and below the
limit node is chosen, which decompresses all the chunks.

This patch ensures that always pathkeys are present for limit queries
and PostgreSQL consideres MergeAppend plans.

Fixes: #4223