Bulk decompression of compressed batches
Decompress each compressed batch entirely in one go, to increase the
throughput. Leads to about 15% speedup on queries that read a lot of
data from a compressed hypertable.

This optimization might increase query run time and memory usage in
cases that have a small LIMIT and/or perform a sorted merge of
compressed batches. We don't have planner support to decide whether it
should be used, so it comes with a GUC to disable it.