PG16: Align GUC variables initial value with boot values
Any value hardcoded to a GUC variable will be overwritten by the
boot_value when the GUC mechanism starts up. PG16 makes this more clear
by checking that if a hardcoded value exists, it is same as the
boot_value. The server asserts in debug builds if values are not equal.

Commit 09636092 already fixes most of our code to align the hardcoded
and boot_values for various GUC variables. This patch updates only the
ts_guc_max_cached_chunks_per_hypertable initialisation which
had an initial value different from the boot_value.

postgres/postgres@a73952b79